validazioni, messaggio success e messaggi di errore in italiano sul form completati

+++ database/migrations/2024_03_05_194703_create_annuncements_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('body');
-            $table->decimal('price',8,2);
+            $table->float('price', 8,2);
             $table->timestamps();
         });
     }
